consensus (refactor): new shared interface of consensus for pacemaker

+++ consensus/e2e_tests/pacemaker_test.go
@@ -141,7 +141,7 @@ func TestPacemakerCatchupSameStepDifferentRounds(t *testing.T) {
 
 	// Starting point
 	testHeight := uint64(3)
-	testStep := consensus.NewRound
+	testStep := uint64(consensus.NewRound)
 
 	// UnitTestNet configs
 	paceMakerTimeoutMsec := uint64(500) // Set a small pacemaker timeout
